+++ Changelog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.12.2] - 2023/01/20
          - [DCAEGEN2-3334] - Update tests execution to be platform agnostic
+         - [DCAEGEN2-3345] - Add additional tests for ConfigProcessor and ApiAuthInterceptor
          
 ## [1.12.1] - 2022/12/05
          - [DCAEGEN2-3257] - Align DCAE components with the new logging GR.